# Final_Project-OpenAPI.yaml
openapi: 3.0.0
info:
  title: FINAL PROJECT
  contact: {}
  version: "1.0"
servers:
  - url: http://localhost:8000/v1/api
    variables: {}
  - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api
    variables: {}
paths:
...
  /airports:
    get:
      tags:
        - AIRPORT FOR USERS
      summary: GET AIRPORT ALL
      operationId: GETAIRPORTALL
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
    post:
      tags:
        - CRUD AIRPORT
      summary: CREATE AIRPORT
      operationId: CREATEAIRPORT
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CREATEAIRPORTREQUEST"
                - example:
                    name: El Mellah Airport
                    code: AAE
                    city: El Tarf
                    country: Algeria
                    terminal: Annaba
                    status: true
            example:
              name: El Mellah Airport
              code: AAE
              city: El Tarf
              country: Algeria
              terminal: Annaba
              status: true
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
...
  /airports/byid/{id}:
    get:
      tags:
        - AIRPORT FOR USERS
      summary: GET AIRPORT by ID
      operationId: GETAIRPORTbyID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
...
  /history/delete/{id}:
    delete:
      tags:
        - HISTORY FOR USERS
      summary: DELETE HISTORY BY ID
      operationId: DELETEHISTORYBYID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
...
  /flight/byid/{id}:
    get:
      tags:
        - CRUD FLIGHT
      summary: GET FLIGHT BY ID
      operationId: GETFLIGHTBYID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
  /flight/create:
    post:
      tags:
        - CRUD FLIGHT
      summary: CREATE FLIGHT
      operationId: CREATEFLIGHT
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CREATEFLIGHTREQUEST"
                - example:
                    flight:
                      departureAirport: 1
                      arrivalAirport: 2
                      departureDate: 2022-09-30
                      arrivalDate: 2022-10-02
                      departureTime: 23:05:59
                      arrivalTime: 07:09:20
                      flightType: 1
                      planeId: 1
                    ticket:
                      class_id: 1
                      price: 1231412
                      country: Indonesia
            example:
              flight:
                departureAirport: 1
                arrivalAirport: 2
                departureDate: 2022-09-30
                arrivalDate: 2022-10-02
                departureTime: 23:05:59
                arrivalTime: 07:09:20
                flightType: 1
                planeId: 1
              ticket:
                class_id: 1
                price: 1231412
                country: Indonesia
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
...
  /flight/delete/{id}:
    delete:
      tags:
        - CRUD FLIGHT
      summary: DELETE FLIGHT
      operationId: DELETEFLIGHT
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://final-Project-FSW1-AND1.up.railway.app/v1/api/
          variables: {}
...
components:
  schemas:
...
    CREATEAIRPORTREQUEST:
      title: CREATEAIRPORTREQUEST
      required:
        - name
        - code
        - city
        - country
        - terminal
        - status
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        city:
          type: string
        country:
          type: string
        terminal:
          type: string
        status:
          type: boolean
      example:
        name: El Mellah Airport
        code: AAE
        city: El Tarf
        country: Algeria
        terminal: Annaba
        status: true
...
    CREATEFLIGHTREQUEST:
      title: CREATEFLIGHTREQUEST
      required:
        - flight
        - ticket
      type: object
      properties:
        flight:
          $ref: "#/components/schemas/Flight"
        ticket:
          $ref: "#/components/schemas/Ticket1"
      example:
        flight:
          departureAirport: 1
          arrivalAirport: 2
          departureDate: 2022-09-30
          arrivalDate: 2022-10-02
          departureTime: 23:05:59
          arrivalTime: 07:09:20
          flightType: 1
          planeId: 1
        ticket:
          class_id: 1
          price: 1231412
          country: Indonesia
    Flight:
      title: Flight
      required:
        - departureAirport
        - arrivalAirport
        - departureDate
        - arrivalDate
        - departureTime
        - arrivalTime
        - flightType
        - planeId
      type: object
      properties:
        departureAirport:
          type: integer
          format: int32
        arrivalAirport:
          type: integer
          format: int32
        departureDate:
          type: string
        arrivalDate:
          type: string
        departureTime:
          type: string
        arrivalTime:
          type: string
        flightType:
          type: integer
          format: int32
        planeId:
          type: integer
          format: int32
      example:
        departureAirport: 1
        arrivalAirport: 2
        departureDate: 2022-09-30
        arrivalDate: 2022-10-02
        departureTime: 23:05:59
        arrivalTime: 07:09:20
        flightType: 1
        planeId: 1
    Ticket1:
      title: Ticket1
      required:
        - class_id
        - price
        - country
      type: object
      properties:
        class_id:
          type: integer
          format: int32
        price:
          type: integer
          format: int32
        country:
          type: string
      example:
        class_id: 1
        price: 1231412
        country: Indonesia
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...